Remove JMediaFileService
