+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/upnp/processor/RecentAlbumId3UpnpProcessor.java
@@ -27,7 +27,7 @@ import javax.annotation.PostConstruct;
 import com.tesshu.jpsonic.dao.AlbumDao;
 import com.tesshu.jpsonic.domain.Album;
 import com.tesshu.jpsonic.domain.logic.CoverArtLogic;
-import com.tesshu.jpsonic.service.JMediaFileService;
+import com.tesshu.jpsonic.service.MediaFileService;
 import com.tesshu.jpsonic.service.upnp.UpnpProcessDispatcher;
 import org.fourthline.cling.support.model.BrowseResult;
 import org.fourthline.cling.support.model.DIDLContent;
@@ -43,7 +43,7 @@ public class RecentAlbumId3UpnpProcessor extends AlbumUpnpProcessor {
     private final UpnpProcessorUtil util;
     private final AlbumDao albumDao;
 
-    public RecentAlbumId3UpnpProcessor(@Lazy UpnpProcessDispatcher d, UpnpProcessorUtil u, JMediaFileService m,
+    public RecentAlbumId3UpnpProcessor(@Lazy UpnpProcessDispatcher d, UpnpProcessorUtil u, MediaFileService m,
             AlbumDao a, CoverArtLogic c) {
         super(d, u, m, a, c);
         this.util = u;
